TIP-860: Move email recorder from enrich bundle to import export bundle
